cvarcmd: add alias command and remove redundant locking mechanisms
the alias could be particularly useful to users who want to chain
together commands, but this is untested. it works for aliasing
variables to more memorable names.

the locking mechanism in CVarCmdService was determined to be redundant
and not being very helpful. adding cvars, cmds, and aliases in multiple
threads is a pretty stupid thing to do anyway...